requirements.txt
Django==1.5.1
psycopg2==2.4.6
python-dateutil==1.5
djangorestframework